Add build output (venus-metrics.min.js) to repo
When building Venus, we don't use npm; and don't want to use it for various
reasons (support of it in OE is bad, and adding it just for this will increase
build time tremendously).

Therefore add the build output to the repo. This does means a change in
workflow when developing.

Alternatively replace npm by something else. Python for example, or C or
something else thats not a complete toolchain like node.